# api/openapi.yaml
openapi: 3.0.4
info:
  title: UnderSound Content API
  description: |
    API de contenido inspirada en Bandcamp para gestionar **álbumes, pistas, merchandising, labels y comentarios**.
    Incluye paginación, filtros, subida de archivos (multipart), acciones de publicación y favoritos.
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/escuelapolitecnicaca/contenido/1.0.0
  description: SwaggerHub API Auto Mocking
- url: https://api.undersound.example.com/v1
  description: Producción
- url: http://localhost:8081
  description: Local
security:
- oauth2:
  - content.read
  - content.write
...
paths:
...
  /albums/{albumId}/cover:
    post:
      tags:
      - Uploads
      - Albums
      summary: Subir o actualizar portada del álbum
      operationId: albumsAlbumIdCoverPOST
      parameters:
      - name: albumId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/albumId_cover_body"
        required: true
      responses:
        "200":
          description: Álbum con portada actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlbumResponse"
      x-swagger-router-controller: Uploads
...
  /tracks:
    get:
      tags:
      - Tracks
      summary: Listar pistas
      operationId: tracksGET
      parameters:
      - name: page
        in: query
        required: false
        style: form
        explode: true
        schema:
          minimum: 1
          type: integer
          default: 1
      - name: limit
        in: query
        required: false
        style: form
        explode: true
        schema:
          maximum: 100
          minimum: 1
          type: integer
          default: 20
      - name: include
        in: query
        description: "Campos relacionados a incluir, separados por coma. Ej. `tracks,label,stats`"
        required: false
        style: form
        explode: false
        schema:
          type: array
          items:
            type: string
      - name: albumId
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          format: uuid
      - name: artistId
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          description: Puede ser UUID o ID numérico del artista
      - name: labelId
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          format: uuid
      - name: genre
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: tag
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: language
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: es
      - name: minDurationSec
        in: query
        required: false
        style: form
        explode: true
        schema:
          minimum: 0
          type: integer
      - name: maxDurationSec
        in: query
        required: false
        style: form
        explode: true
        schema:
          minimum: 0
          type: integer
      - name: releasedFrom
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          format: date
      - name: releasedTo
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          format: date
      - name: sort
        in: query
        description: Campo de ordenación
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - title
          - durationSec
          - playCount
      - name: order
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          default: asc
          enum:
          - asc
          - desc
      - name: q
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: Lista paginada de pistas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTrackList"
      x-swagger-router-controller: Tracks
    post:
      tags:
      - Tracks
      summary: Crear pista
      operationId: tracksPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackCreateInput"
        required: true
      responses:
        "201":
          description: Pista creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackResponse"
      x-swagger-router-controller: Tracks
...
  /tracks/{trackId}/audio:
    post:
      tags:
      - Uploads
      - Tracks
      summary: Subir o actualizar archivo de audio
      operationId: tracksTrackIdAudioPOST
      parameters:
      - name: trackId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/trackId_audio_body"
        required: true
      responses:
        "200":
          description: Pista con audio actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackResponse"
      x-swagger-router-controller: Uploads
...
  /merch:
    get:
      tags:
      - Merch
      summary: Listar productos de merchandising
      operationId: merchGET
      parameters:
      - name: page
        in: query
        required: false
        style: form
        explode: true
        schema:
          minimum: 1
          type: integer
          default: 1
      - name: limit
        in: query
        required: false
        style: form
        explode: true
        schema:
          maximum: 100
          minimum: 1
          type: integer
          default: 20
      - name: artistId
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          description: ID del artista (puede ser número o UUID)
      - name: labelId
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          description: ID del label (puede ser número o UUID)
      - name: type
        in: query
        description: Tipo de producto
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - camiseta
          - vinilo
          - cd
          - poster
          - pegatina
          - otro
      - name: availability
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - in_stock
          - out_of_stock
          - preorder
      - name: sort
        in: query
        description: Campo de ordenación
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - name
          - price
          - createdAt
      - name: order
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          default: asc
          enum:
          - asc
          - desc
      - name: q
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: Lista paginada de merchandising
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedMerchList"
      x-swagger-router-controller: Merch
    post:
      tags:
      - Merch
      summary: Crear producto de merchandising
      operationId: merchPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MerchCreateInput"
        required: true
      responses:
        "201":
          description: Producto creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchResponse"
      x-swagger-router-controller: Merch
...
  /merch/{merchId}/images:
    post:
      tags:
      - Uploads
      - Merch
      summary: Subir imágenes del producto
      operationId: merchMerchIdImagesPOST
      parameters:
      - name: merchId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/merchId_images_body"
        required: true
      responses:
        "200":
          description: Producto con imágenes actualizadas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchResponse"
      x-swagger-router-controller: Uploads
...
components:
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 123
      example:
        total: 123
        limit: 20
        page: 1
    Image:
      type: object
      properties:
        url:
          type: string
          format: uri
        alt:
          type: string
        width:
          type: integer
        height:
          type: integer
      example:
        alt: alt
        width: 6
        url: http://example.com/aeiou
        height: 1
    ArtistRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
      example:
        name: name
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
    LabelRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
      example:
        name: name
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
    Album:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        artist:
          $ref: "#/components/schemas/ArtistRef"
        label:
          $ref: "#/components/schemas/LabelRef"
        releaseDate:
          type: string
          format: date
        releaseState:
          type: string
          enum:
          - draft
          - scheduled
          - published
          - archived
        price:
          type: number
          format: float
        currency:
          type: string
          example: EUR
        genres:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        cover:
          $ref: "#/components/schemas/Image"
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/Track"
        stats:
          $ref: "#/components/schemas/Album_stats"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        artist:
          name: name
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        releaseDate: 2000-01-23
        description: description
        label:
          name: name
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        title: title
        tracks:
        - createdAt: 2000-01-23T04:56:07.000+00:00
          trackNumber: 5
          artist: null
          stats:
            playCount: 7
          album:
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            title: title
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          label: null
          audio:
            codec: codec
            bitrate: 2
            url: http://example.com/aeiou
          title: title
          durationSec: 5
          lyrics:
            language: language
            text: text
          updatedAt: 2000-01-23T04:56:07.000+00:00
        - createdAt: 2000-01-23T04:56:07.000+00:00
          trackNumber: 5
          artist: null
          stats:
            playCount: 7
          album:
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            title: title
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          label: null
          audio:
            codec: codec
            bitrate: 2
            url: http://example.com/aeiou
          title: title
          durationSec: 5
          lyrics:
            language: language
            text: text
          updatedAt: 2000-01-23T04:56:07.000+00:00
        tags:
        - tags
        - tags
        cover:
          alt: alt
          width: 6
          url: http://example.com/aeiou
          height: 1
        createdAt: 2000-01-23T04:56:07.000+00:00
        stats:
          playCount: 9
          likeCount: 3
          ratingAvg: 4.5
          commentCount: 2
        price: 0.8008282
        genres:
        - genres
        - genres
        currency: EUR
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        releaseState: draft
        updatedAt: 2000-01-23T04:56:07.000+00:00
...
    AlbumResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Album"
      example:
        data:
          artist:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          releaseDate: 2000-01-23
          description: description
          label:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          title: title
          tracks:
          - createdAt: 2000-01-23T04:56:07.000+00:00
            trackNumber: 5
            artist: null
            stats:
              playCount: 7
            album:
              id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
              title: title
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            label: null
            audio:
              codec: codec
              bitrate: 2
              url: http://example.com/aeiou
            title: title
            durationSec: 5
            lyrics:
              language: language
              text: text
            updatedAt: 2000-01-23T04:56:07.000+00:00
          - createdAt: 2000-01-23T04:56:07.000+00:00
            trackNumber: 5
            artist: null
            stats:
              playCount: 7
            album:
              id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
              title: title
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            label: null
            audio:
              codec: codec
              bitrate: 2
              url: http://example.com/aeiou
            title: title
            durationSec: 5
            lyrics:
              language: language
              text: text
            updatedAt: 2000-01-23T04:56:07.000+00:00
          tags:
          - tags
          - tags
          cover:
            alt: alt
            width: 6
            url: http://example.com/aeiou
            height: 1
          createdAt: 2000-01-23T04:56:07.000+00:00
          stats:
            playCount: 9
            likeCount: 3
            ratingAvg: 4.5
            commentCount: 2
          price: 0.8008282
          genres:
          - genres
          - genres
          currency: EUR
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          releaseState: draft
          updatedAt: 2000-01-23T04:56:07.000+00:00
...
    Track:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        durationSec:
          type: integer
        trackNumber:
          type: integer
        artist:
          $ref: "#/components/schemas/ArtistRef"
        album:
          $ref: "#/components/schemas/AlbumRef"
        label:
          $ref: "#/components/schemas/LabelRef"
        audio:
          $ref: "#/components/schemas/Track_audio"
        lyrics:
          $ref: "#/components/schemas/Track_lyrics"
        stats:
          $ref: "#/components/schemas/Track_stats"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        trackNumber: 5
        artist: null
        stats:
          playCount: 7
        album:
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          title: title
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        label: null
        audio:
          codec: codec
          bitrate: 2
          url: http://example.com/aeiou
        title: title
        durationSec: 5
        lyrics:
          language: language
          text: text
        updatedAt: 2000-01-23T04:56:07.000+00:00
    AlbumRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
      example:
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        title: title
    TrackCreateMinimal:
      required:
      - title
      type: object
      properties:
        title:
          type: string
        durationSec:
          type: integer
        artistId:
          type: string
          format: uuid
    TrackCreateInput:
      allOf:
      - $ref: "#/components/schemas/TrackCreateMinimal"
      - type: object
        properties:
          albumId:
            type: string
            format: uuid
          labelId:
            type: string
            format: uuid
          trackNumber:
            type: integer
          tags:
            type: array
            items:
              type: string
...
    PaginatedTrackList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Track"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      example:
        data:
        - createdAt: 2000-01-23T04:56:07.000+00:00
          trackNumber: 5
          artist: null
          stats:
            playCount: 7
          album:
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            title: title
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          label: null
          audio:
            codec: codec
            bitrate: 2
            url: http://example.com/aeiou
          title: title
          durationSec: 5
          lyrics:
            language: language
            text: text
          updatedAt: 2000-01-23T04:56:07.000+00:00
        - createdAt: 2000-01-23T04:56:07.000+00:00
          trackNumber: 5
          artist: null
          stats:
            playCount: 7
          album:
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            title: title
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          label: null
          audio:
            codec: codec
            bitrate: 2
            url: http://example.com/aeiou
          title: title
          durationSec: 5
          lyrics:
            language: language
            text: text
          updatedAt: 2000-01-23T04:56:07.000+00:00
        meta:
          total: 123
          limit: 20
          page: 1
    TrackResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Track"
      example:
        data:
          createdAt: 2000-01-23T04:56:07.000+00:00
          trackNumber: 5
          artist: null
          stats:
            playCount: 7
          album:
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            title: title
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          label: null
          audio:
            codec: codec
            bitrate: 2
            url: http://example.com/aeiou
          title: title
          durationSec: 5
          lyrics:
            language: language
            text: text
          updatedAt: 2000-01-23T04:56:07.000+00:00
...
    MerchItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum:
          - camiseta
          - vinilo
          - cd
          - poster
          - pegatina
          - otro
        images:
          type: array
          items:
            $ref: "#/components/schemas/Image"
        variants:
          type: array
          items:
            $ref: "#/components/schemas/Variant"
        label:
          $ref: "#/components/schemas/LabelRef"
        artist:
          $ref: "#/components/schemas/ArtistRef"
        availability:
          type: string
          enum:
          - in_stock
          - out_of_stock
          - preorder
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        images:
        - alt: alt
          width: 6
          url: http://example.com/aeiou
          height: 1
        - alt: alt
          width: 6
          url: http://example.com/aeiou
          height: 1
        artist:
          name: name
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        name: name
        description: description
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        variants:
        - color: color
          size: size
          price: 0.8008282
          currency: EUR
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          sku: sku
          stock: 6
        - color: color
          size: size
          price: 0.8008282
          currency: EUR
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          sku: sku
          stock: 6
        label:
          name: name
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        availability: in_stock
        type: camiseta
        updatedAt: 2000-01-23T04:56:07.000+00:00
...
    MerchCreateInput:
      required:
      - name
      - type
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum:
          - camiseta
          - vinilo
          - cd
          - poster
          - pegatina
          - otro
        artistId:
          type: string
          description: ID del artista (usuario autenticado) - puede ser número o UUID
        labelId:
          type: string
          description: ID del sello discográfico (opcional) - puede ser número o UUID
...
    PaginatedMerchList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MerchItem"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      example:
        data:
        - createdAt: 2000-01-23T04:56:07.000+00:00
          images:
          - alt: alt
            width: 6
            url: http://example.com/aeiou
            height: 1
          - alt: alt
            width: 6
            url: http://example.com/aeiou
            height: 1
          artist:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          name: name
          description: description
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          variants:
          - color: color
            size: size
            price: 0.8008282
            currency: EUR
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            sku: sku
            stock: 6
          - color: color
            size: size
            price: 0.8008282
            currency: EUR
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            sku: sku
            stock: 6
          label:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          availability: in_stock
          type: camiseta
          updatedAt: 2000-01-23T04:56:07.000+00:00
        - createdAt: 2000-01-23T04:56:07.000+00:00
          images:
          - alt: alt
            width: 6
            url: http://example.com/aeiou
            height: 1
          - alt: alt
            width: 6
            url: http://example.com/aeiou
            height: 1
          artist:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          name: name
          description: description
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          variants:
          - color: color
            size: size
            price: 0.8008282
            currency: EUR
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            sku: sku
            stock: 6
          - color: color
            size: size
            price: 0.8008282
            currency: EUR
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            sku: sku
            stock: 6
          label:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          availability: in_stock
          type: camiseta
          updatedAt: 2000-01-23T04:56:07.000+00:00
        meta:
          total: 123
          limit: 20
          page: 1
    MerchResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/MerchItem"
      example:
        data:
          createdAt: 2000-01-23T04:56:07.000+00:00
          images:
          - alt: alt
            width: 6
            url: http://example.com/aeiou
            height: 1
          - alt: alt
            width: 6
            url: http://example.com/aeiou
            height: 1
          artist:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          name: name
          description: description
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          variants:
          - color: color
            size: size
            price: 0.8008282
            currency: EUR
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            sku: sku
            stock: 6
          - color: color
            size: size
            price: 0.8008282
            currency: EUR
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            sku: sku
            stock: 6
          label:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          availability: in_stock
          type: camiseta
          updatedAt: 2000-01-23T04:56:07.000+00:00
...
    albumId_cover_body:
      required:
      - file
      type: object
      properties:
        file:
          type: string
          format: binary
        alt:
          type: string
          description: Texto alternativo
    trackId_audio_body:
      required:
      - file
      type: object
      properties:
        file:
          type: string
          format: binary
          description: Audio file (MP3, WAV, OGG, etc.)
...
    merchId_images_body:
      type: object
      properties:
        files:
          type: array
          items:
            type: string
            format: binary
        alt:
          type: string
...
    Album_stats:
      type: object
      properties:
        playCount:
          type: integer
        likeCount:
          type: integer
        commentCount:
          type: integer
        ratingAvg:
          type: number
          format: float
          example: 4.5
      example:
        playCount: 9
        likeCount: 3
        ratingAvg: 4.5
        commentCount: 2
...
    Track_audio:
      type: object
      properties:
        url:
          type: string
          format: uri
        bitrate:
          type: integer
        codec:
          type: string
      example:
        codec: codec
        bitrate: 2
        url: http://example.com/aeiou
    Track_lyrics:
      type: object
      properties:
        text:
          type: string
        language:
          type: string
      example:
        language: language
        text: text
    Track_stats:
      type: object
      properties:
        playCount:
          type: integer
      example:
        playCount: 7
...
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.undersound.example.com/authorize
          tokenUrl: https://auth.undersound.example.com/token
          scopes:
            content.read: Lectura de contenido
            content.write: Escritura de contenido
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT